grad_project/requirements.txt
asgiref==3.2.10
awscli==1.18.164
boto3==1.16.4
botocore==1.19.4
colorama==0.4.3
Django==3.1.2
djangorestframework==3.12.1
docutils==0.15.2
jmespath==0.10.0
Js2Py==0.70
numpy==1.19.2
Pillow==8.0.1
pyasn1==0.4.8
pygame==1.9.6
pyjsparser==2.7.1
PyOpenGL==3.1.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
rsa==4.5
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.25.11
